feat: Add --update-env flag to update .envrc
This commit introduces a new `--update-env` flag to the `ppg.py` script.  When this flag is used, the script will update the `.envrc` file in the current working directory with the following environment variables:

*   `PPG_OUTPUT_FILE`:  Set to `~/Downloads/$(basename $PWD).txt`
*   `PPG_OUTPUT_DIR`: Set to `~/Downloads/$(basename $PWD)`

The script checks if the lines already exist in the `.envrc` file to avoid duplicates.  This allows users to easily configure the output file and directory for Project Prompt Generator (PPG) via environment variables, and have those variables automatically set when entering a project directory using tools like `direnv`.

A new `prompts/envrc_handler.py` file is created to handle the `.envrc` file update logic.

Signed-off-by: Ching Yi, Chan <chingyichan.tw@gmail.com>